Move resolve_path to vfs::internal
