+++ api/src/vfs/internal.rs
@@ -7,6 +7,8 @@ use crate::{
     vfs::{NodeMetadata, NodePermissions, NodeType},
 };
 use diesel::PgConnection;
+use lazy_static::lazy_static;
+use regex::Regex;
 use std::{collections::HashMap, fmt::Debug};
 
 /// A definition for a path segment in the virtual tree. The definition should
@@ -235,5 +237,63 @@ impl VirtualNode {
                 Ok(rv)
             }
         }
+    }
+}
+
+/// Normalizes the given path, then splits it into segments. Normalization
+/// includes de-duplicating slashes and removing trailing slashes.
+///
+/// In the future, if we want to start supporting relative paths, this could
+/// resolve `.` and `..` as well.
+///
+/// ```
+/// assert_eq!(split_path("/").as_slice(), &[""]);
+/// assert_eq!(split_path("/dir1").as_slice(), &["", "dir1"]);
+/// assert_eq!(split_path("/dir1/dir2/").as_slice(), &["", "dir1", "dir2"]);
+/// ```
+///
+/// See unit tests for more test cases.
+pub fn resolve_path(path: &str) -> Vec<&str> {
+    lazy_static! {
+        static ref PATH_SEP_RGX: Regex = Regex::new(r"/+").unwrap();
+    }
+    match path {
+        // Special case, so that an empty path matches the root
+        "" => vec![""],
+        _ => PATH_SEP_RGX.split(path).collect(),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_resolve_path() {
+        assert_eq!(resolve_path("").as_slice(), &[""]);
+        assert_eq!(resolve_path("/").as_slice(), &[""]);
+        assert_eq!(resolve_path("//").as_slice(), &[""]);
+        assert_eq!(resolve_path("dir1").as_slice(), &["dir1"]);
+        assert_eq!(resolve_path("/dir1").as_slice(), &["", "dir1"]);
+        assert_eq!(
+            resolve_path("/dir1/dir2").as_slice(),
+            &["", "dir1", "dir2"]
+        );
+        assert_eq!(
+            resolve_path("/dir1///dir2/").as_slice(),
+            &["", "dir1", "dir2"]
+        );
+        assert_eq!(
+            resolve_path("/dir1/dir2/file.txt").as_slice(),
+            &["", "dir1", "dir2", "file.txt"]
+        );
+        assert_eq!(
+            resolve_path("/dir1/./file.txt").as_slice(),
+            &["", "dir1", ".", "file.txt"]
+        );
+        assert_eq!(
+            resolve_path("/dir1/../file.txt").as_slice(),
+            &["", "dir1", "..", "file.txt"]
+        );
     }
 }
